move code to remote_machine
