+++ src/nni_manager/training_service/remote_machine/extends/linuxCommands.ts
