+++ src/nni_manager/training_service/remote_machine/osCommands.ts
